1. support_mode num changed to one; 2. i2c read/write reg data size changed to 32bit

+++ drv/fpga_vir_cam.h
@@ -9,7 +9,7 @@
 
 struct fpga_vir_cam_reg {
     unsigned short reg_value;
-    unsigned short reg_addr;
+    unsigned int reg_addr;
 };
